Changed GP to decimal type

+++ Models/Character.cs
@@ -42,7 +42,7 @@ namespace dnd_buddy_backend.Models
 
         public int Xp { get; set; }
 
-        public float Gp { get; set; }
+        public Decimal Gp { get; set; }
 
         public int UserId { get; set; }
